Restructure ledger for mantle ops
Restructure ledger to support mantle operations
