+++ ledger/nomos-ledger/Cargo.toml
@@ -8,7 +8,6 @@ version = "0.1.0"
 workspace = true
 
 [dependencies]
-blake2                 = "0.10"
 cryptarchia-engine     = { workspace = true }
 crypto-bigint          = "0.5"
 nomos-core             = { workspace = true }
