--- ledger/nomos-ledger/src/crypto.rs
@@ -1,3 +1,0 @@
-use blake2::digest::typenum::U32;
-
-pub type Blake2b = blake2::Blake2b<U32>;
