+++ nodes/nomos-executor/config.yaml
@@ -162,49 +162,12 @@ cryptarchia:
       active_slot_coeff: 0.9
   genesis_id: 0x0000000000000000000000000000000000000000000000000000000000000000
   genesis_state:
-    utxos:
+    cryptarchia_ledger:
       utxos:
-        03b2d3b8a925adcb25eb686b92db7f21a75f0e0f7c93da95b383f0fd35279e48:
-          value: 1
-          pk: '0000000000000000000000000000000000000000000000000000000000000000'
-    nonce:
-      [
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-        0,
-      ]
-    slot: 0
-    next_epoch_state:
-      epoch: 1
+        utxos:
+          03b2d3b8a925adcb25eb686b92db7f21a75f0e0f7c93da95b383f0fd35279e48:
+            value: 1
+            pk: '0000000000000000000000000000000000000000000000000000000000000000'
       nonce:
         [
           0,
@@ -240,55 +203,94 @@ cryptarchia:
           0,
           0,
         ]
-      utxos:
+      slot: 0
+      next_epoch_state:
+        epoch: 1
+        nonce:
+          [
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+          ]
         utxos:
-          03b2d3b8a925adcb25eb686b92db7f21a75f0e0f7c93da95b383f0fd35279e48:
-            value: 1
-            pk: '0000000000000000000000000000000000000000000000000000000000000000'
-      total_stake: 1
-    epoch_state:
-      epoch: 0
-      nonce:
-        [
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-          0,
-        ]
-      utxos:
+          utxos:
+            03b2d3b8a925adcb25eb686b92db7f21a75f0e0f7c93da95b383f0fd35279e48:
+              value: 1
+              pk: '0000000000000000000000000000000000000000000000000000000000000000'
+        total_stake: 1
+      epoch_state:
+        epoch: 0
+        nonce:
+          [
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+            0,
+          ]
         utxos:
-          03b2d3b8a925adcb25eb686b92db7f21a75f0e0f7c93da95b383f0fd35279e48:
-            value: 1
-            pk: '0000000000000000000000000000000000000000000000000000000000000000'
-      total_stake: 1
+          utxos:
+            03b2d3b8a925adcb25eb686b92db7f21a75f0e0f7c93da95b383f0fd35279e48:
+              value: 1
+              pk: '0000000000000000000000000000000000000000000000000000000000000000'
+        total_stake: 1
+    mantle_ledger: ~
   time:
     slot_duration: [2, 0]
     chain_start_time: [2025, 41, 9, 38, 29, 993653000, 0, 0, 0]
